+++ source/B2B.Transactions.Api/TransactionQueueListener.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Text;
 using System.Threading.Tasks;
+using B2B.Transactions.Configuration;
 using B2B.Transactions.Infrastructure.Configuration.Correlation;
 using B2B.Transactions.Infrastructure.Serialization;
 using B2B.Transactions.Transactions;
