+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -18,6 +18,7 @@ using B2B.CimMessageAdapter;
 using B2B.CimMessageAdapter.Messages;
 using B2B.CimMessageAdapter.Transactions;
 using B2B.Transactions.Authentication;
+using B2B.Transactions.Configuration;
 using B2B.Transactions.DataAccess;
 using B2B.Transactions.Infrastructure.Authentication.Bearer;
 using B2B.Transactions.Infrastructure.Authentication.MarketActors;
